refactor(frontend): share stat card component
