+++ frontend/src/components/common/StatCard.jsx
@@ -1,112 +1,137 @@
 import { Box, Paper, Stack, Typography } from '@mui/material'
 import { alpha, useTheme } from '@mui/material/styles'
-import { getSurfaceGeometry } from './surfaceGeometry.js'
-
-const toneToPaletteKey = {
-  primary: 'primary',
-  secondary: 'secondary',
-  success: 'success',
-  warning: 'warning',
-  info: 'info',
-  error: 'error',
-}
 
 /**
- * Canonical KPI card. A theme-driven, copy-agnostic stat tile used across
- * dashboards. `tone` selects a palette accent; `icon` is optional.
- *
- * Props: label, value, helper?, tone? ('primary'|'secondary'|'success'|'warning'|'info'|'error'), icon?
+ * Ascuas KPI card: label + big value + optional sub/trend line + a muted
+ * corner icon. Mirrors the Overview datacard look (issue #144) so the migrated
+ * Accountability view reads as the same system.
  */
 export default function StatCard({
-  label = '',
-  value = '',
-  helper = '',
-  tone = 'primary',
-  icon = null,
+  label,
+  value,
+  sub,
+  subTone = 'neutral',
+  icon,
+  accent,
+  onClick,
 }) {
   const theme = useTheme()
   const isDark = theme.palette.mode === 'dark'
-  const geometry = getSurfaceGeometry(theme)
-  const paletteKey = toneToPaletteKey[tone] || 'primary'
-  const accent = theme.palette[paletteKey].main
-  const valueText = String(value ?? '').trim() || '-'
+  const radius = theme.custom?.dashboard?.radius?.surface || '14px'
+  const accentColor = accent || theme.palette.secondary.main
+
+  const toneColor =
+    subTone === 'positive'
+      ? theme.palette.success.main
+      : subTone === 'negative'
+        ? theme.palette.error.main
+        : '#88736A'
+
+  const interactive = typeof onClick === 'function'
 
   return (
     <Paper
       elevation={0}
+      onClick={onClick}
+      {...(interactive
+        ? {
+            role: 'button',
+            tabIndex: 0,
+            onKeyDown: (event) => {
+              if (event.key === 'Enter' || event.key === ' ') {
+                event.preventDefault()
+                onClick()
+              }
+            },
+          }
+        : {})}
       sx={{
-        minHeight: '100%',
-        borderRadius: geometry.surfaceRadius,
         position: 'relative',
         overflow: 'hidden',
-        border: `1px solid ${alpha(theme.palette.text.primary, geometry.subtleBorderAlpha)}`,
-        background: `linear-gradient(180deg, ${alpha(theme.palette.background.paper, 0.98)} 0%, ${alpha(
-          theme.palette.background.default,
-          isDark ? 0.72 : 0.7
-        )} 100%)`,
-        boxShadow: geometry.cardShadow,
-        '&::before': {
-          content: '""',
-          position: 'absolute',
-          inset: 0,
-          background: `linear-gradient(145deg, ${alpha(accent, 0.06)} 0%, transparent 42%)`,
-          pointerEvents: 'none',
-        },
+        minHeight: 118,
+        height: '100%',
+        display: 'flex',
+        alignItems: 'center',
+        borderRadius: radius,
+        border: `1px solid ${alpha(theme.palette.text.primary, isDark ? 0.05 : 0.07)}`,
+        backgroundColor: '#45342C',
+        px: '24px',
+        py: '18px',
+        ...(interactive
+          ? {
+              cursor: 'pointer',
+              transition: 'transform 150ms ease, border-color 150ms ease',
+              '&:hover': {
+                transform: 'translateY(-2px)',
+                borderColor: alpha(accentColor, 0.6),
+              },
+              '&.Mui-focusVisible, &:focus-visible': {
+                outline: `2px solid ${alpha(accentColor, 0.9)}`,
+                outlineOffset: 2,
+              },
+            }
+          : {}),
       }}
     >
-      <Box sx={{ position: 'relative', zIndex: 1, px: 1.5, py: 1.25 }}>
-        <Stack spacing={0.65} sx={{ minWidth: 0 }}>
-          <Stack direction="row" alignItems="center" justifyContent="space-between" spacing={1}>
-            <Typography
-              variant="overline"
-              color="text.secondary"
-              sx={{ letterSpacing: 0.5, lineHeight: 1.05, maxWidth: '78%' }}
-            >
-              {label}
-            </Typography>
-            {icon ? (
-              <Box
-                sx={{
-                  width: 24,
-                  height: 24,
-                  flexShrink: 0,
-                  borderRadius: geometry.badgeRadius,
-                  display: 'grid',
-                  placeItems: 'center',
-                  color: accent,
-                  bgcolor: alpha(accent, 0.1),
-                  border: `1px solid ${alpha(accent, 0.14)}`,
-                }}
-              >
-                {icon}
-              </Box>
-            ) : null}
-          </Stack>
+      {icon ? (
+        <Box
+          component="span"
+          className="material-symbols-rounded"
+          sx={{
+            position: 'absolute',
+            right: 16,
+            bottom: 12,
+            fontSize: 44,
+            color: alpha(accentColor, 0.5),
+            pointerEvents: 'none',
+            fontVariationSettings: "'FILL' 0, 'wght' 400, 'GRAD' 0, 'opsz' 40",
+          }}
+        >
+          {icon}
+        </Box>
+      ) : null}
 
+      <Stack spacing="6px" sx={{ position: 'relative', zIndex: 1, pr: 4, width: '100%' }}>
+        <Typography
+          sx={{
+            fontFamily: '"JetBrains Mono", monospace',
+            fontWeight: 600,
+            textTransform: 'uppercase',
+            letterSpacing: '0.06em',
+            fontSize: '0.82rem',
+            lineHeight: 1.25,
+            color: '#88736A',
+          }}
+        >
+          {label}
+        </Typography>
+        <Typography
+          sx={{
+            fontFamily: '"Hanken Grotesk", sans-serif',
+            fontWeight: 800,
+            color: '#F4EEE8',
+            lineHeight: 1.02,
+            fontSize: '2.1rem',
+            letterSpacing: -0.5,
+            overflowWrap: 'anywhere',
+          }}
+        >
+          {value}
+        </Typography>
+        {sub ? (
           <Typography
             sx={{
-              fontWeight: 700,
-              color: 'text.primary',
-              lineHeight: 1.08,
-              fontSize: '1.34rem',
-              letterSpacing: -0.2,
-              overflowWrap: 'anywhere',
+              fontFamily: '"JetBrains Mono", monospace',
+              fontWeight: 600,
+              fontSize: '0.72rem',
+              letterSpacing: '0.02em',
+              color: toneColor,
             }}
           >
-            {valueText}
+            {sub}
           </Typography>
-
-          {helper ? (
-            <Typography
-              variant="body2"
-              color="text.secondary"
-              sx={{ fontSize: '0.76rem', lineHeight: 1.25 }}
-            >
-              {helper}
-            </Typography>
-          ) : null}
-        </Stack>
-      </Box>
+        ) : null}
+      </Stack>
     </Paper>
   )
 }
